+++ packages/proto-build/src/transformers.rs
@@ -1,0 +1,186 @@
+use crate::{format_ident, quote};
+use heck::ToUpperCamelCase;
+use prost_types::{FileDescriptorSet, ServiceDescriptorProto};
+use std::collections::HashMap;
+use std::path::Path;
+use syn::{Attribute, Fields, Ident, Item, parse_quote, Type};
+use regex::Regex;
+use syn::__private::quote::__private::TokenStream as TokenStream2;
+use heck::ToSnakeCase;
+
+/// Regex substitutions to apply to the prost-generated output
+pub const REPLACEMENTS: &[(&str, &str)] = &[
+    // Use `tendermint-proto` proto definitions
+    ("(super::)+tendermint", "tendermint_proto"),
+    // Feature-gate gRPC client modules
+    (
+        "/// Generated client implementations.",
+        "/// Generated client implementations.\n\
+             #[cfg(feature = \"grpc\")]\n\
+             #[cfg_attr(docsrs, doc(cfg(feature = \"grpc\")))]",
+    ),
+    // Feature-gate gRPC client impls which use `tonic::transport`
+    (
+        "impl (.+)Client<tonic::transport::Channel>",
+        "#[cfg(feature = \"grpc-transport\")]\n    \
+             #[cfg_attr(docsrs, doc(cfg(feature = \"grpc-transport\")))]\n    \
+             impl ${1}Client<tonic::transport::Channel>",
+    ),
+];
+
+pub fn append_attrs(
+    src: &Path,
+    ancestors: &[String],
+    ident: &Ident,
+    attrs: &mut Vec<Attribute>,
+    descriptor: &FileDescriptorSet,
+) {
+    let query_services = extract_query_services(descriptor);
+    let type_url = get_type_url(src, ident);
+    let path: Vec<String> = type_url.split('.').map(|s| s.to_string()).collect();
+    let type_url = [&path[..(path.len() - 1)], ancestors, &[ident.to_string()]]
+        .concat()
+        .join(".");
+    attrs.append(&mut vec![
+        syn::parse_quote! { #[derive(serde::Serialize, serde::Deserialize, schemars::JsonSchema, CosmwasmExt)] },
+        syn::parse_quote! { #[proto_message(type_url = #type_url)] },
+    ]);
+
+    if let Some(attr) = get_query_attr(src, ident, &query_services) {
+        attrs.append(&mut vec![attr])
+    }
+}
+
+// ====== helpers ======
+
+fn get_query_attr(
+    src: &Path,
+    ident: &Ident,
+    query_services: &HashMap<String, ServiceDescriptorProto>,
+) -> Option<Attribute> {
+    let package = src.file_stem().unwrap().to_str().unwrap();
+    let service = query_services.get(package);
+
+    let method = service?.method.iter().find(|m| {
+        let input_type = (*m).input_type.clone().unwrap();
+        let input_type = input_type.split('.').last().unwrap();
+        *ident == input_type.to_upper_camel_case()
+    });
+
+    let method_name = method?.name.clone().unwrap();
+    let response_type = method?.output_type.clone().unwrap();
+    let response_type = response_type.split('.').last().unwrap();
+    let response_type = format_ident!("{}", response_type.to_upper_camel_case());
+
+    let path = format!("/{}.Query/{}", package, method_name);
+    Some(syn::parse_quote! { #[proto_query(path = #path, response_type = #response_type)] })
+}
+
+fn get_type_url(src: &Path, ident: &Ident) -> String {
+    let type_path = src.file_stem().unwrap().to_str().unwrap();
+    format!("/{}.{}", type_path, ident)
+}
+
+pub fn extract_query_services(
+    descriptor: &FileDescriptorSet,
+) -> HashMap<String, ServiceDescriptorProto> {
+    descriptor
+        .clone()
+        .file
+        .into_iter()
+        .filter_map(|f| {
+            let service = f
+                .service
+                .into_iter()
+                .find(|s| s.name == Some("Query".to_string()));
+
+            if let Some(service) = service {
+                Some((
+                    f.package.expect("Missing package name in file descriptor"),
+                    service,
+                ))
+            } else {
+                None
+            }
+        })
+        .collect()
+}
+
+pub fn append_querier(
+    items: Vec<Item>,
+    src: &Path,
+    nested_mod: bool,
+    descriptor: &FileDescriptorSet,
+) -> Vec<Item> {
+    let package = src.file_stem().unwrap().to_str().unwrap();
+    let re = Regex::new(r"([^.]*)(\.v\d+(beta\d+)?)?$").unwrap();
+
+    let package_stem = re.captures(&package).unwrap().get(1).unwrap().as_str();
+
+    let querier_wrapper_ident = format_ident!("{}Querier", &package_stem.to_upper_camel_case());
+
+    let query_services = extract_query_services(descriptor);
+    let query_fns = query_services.get(package).map(|service| service.method.iter().map(|method_desc| {
+        if nested_mod {
+            return quote! {};
+        }
+
+        let method_desc = method_desc.clone();
+
+        let name = format_ident!("{}", method_desc.name.unwrap().as_str().to_snake_case());
+        let req_type = format_ident!("{}", method_desc.input_type.unwrap().split('.').last().unwrap().to_string().to_upper_camel_case());
+        let res_type = format_ident!("{}", method_desc.output_type.unwrap().split('.').last().unwrap().to_string().to_upper_camel_case());
+
+        let req_args = items.clone().into_iter()
+            .find_map(|item| match item {
+                Item::Struct(s) => {
+                    if s.ident == req_type {
+                        match s.fields {
+                            Fields::Named(fields_named) => {
+                                Some(fields_named.named)
+                            }
+                            _ => None
+                        }
+                    } else {
+                        None
+                    }
+                }
+                _ => None
+            });
+
+        let arg_idents = req_args.clone().unwrap().into_iter().map(|arg| arg.ident.unwrap()).collect::<Vec<Ident>>();
+        let arg_ty = req_args.clone().unwrap().into_iter().map(|arg| arg.ty).collect::<Vec<Type>>();
+
+        quote! {
+          pub fn #name( &self, #(#arg_idents : #arg_ty),* ) -> Result<#res_type, cosmwasm_std::StdError> {
+            #req_type { #(#arg_idents),* }.query(self.querier)
+          }
+        }
+    }).collect::<Vec<TokenStream2>>());
+
+    let querier = if let Some(query_fns) = query_fns {
+        if !nested_mod {
+            vec![
+                parse_quote! {
+                  pub struct #querier_wrapper_ident<'a> {
+                    querier: cosmwasm_std::QuerierWrapper<'a, cosmwasm_std::Empty>
+                  }
+                },
+                parse_quote! {
+                  impl<'a> #querier_wrapper_ident<'a> {
+                    pub fn new(querier: cosmwasm_std::QuerierWrapper<'a, cosmwasm_std::Empty>) -> Self {
+                      Self { querier }
+                    }
+                    #(#query_fns)*
+                  }
+                },
+            ]
+        } else {
+            vec![]
+        }
+    } else {
+        vec![]
+    };
+
+    vec![items, querier].concat()
+}
